+++ src/4-executing/builtins/util/envcp_utils.c
@@ -6,7 +6,7 @@
 /*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/21 07:17:01 by fschuber          #+#    #+#             */
-/*   Updated: 2024/04/02 10:09:52 by fschuber         ###   ########.fr       */
+/*   Updated: 2024/04/03 09:50:06 by fschuber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,30 +17,29 @@
 // so strdup it if you want to keep it
 char	*get_envcp_var(char *var, char **envcp)
 {
-	int	env_i;
-	int	env_i_i;
+	int	i;
+	int	j;
 	int	var_i;
 
-	env_i = 0;
-	while (envcp[env_i])
+	i = 0;
+	while (envcp[i])
 	{
-		env_i_i = 0;
+		j = 0;
 		var_i = 0;
-		while (envcp[env_i][env_i_i] && var[var_i] && envcp[env_i][env_i_i] == var[var_i])
+		while (envcp[i][j] && var[var_i] && envcp[i][j] == var[var_i])
 		{
-			env_i_i++;
+			j++;
 			var_i++;
 		}
-		if (var[var_i] == '\0' && envcp[env_i][env_i_i] == '=')
-			return (envcp[env_i] + env_i_i + 1);
-		env_i++;
+		if (var[var_i] == '\0' && envcp[i][j] == '=')
+			return (envcp[i] + j + 1);
+		i++;
 	}
 	return (NULL);
 }
 
 // creates a new environment variable with the given value
-int	create_envcp_var(char *var, char *value, char **envcp, \
-					t_program_data *program_data)
+int	create_envcp_var(char *var, char *value, t_program_data *data)
 {
 	char	*newenvcp;
 	char	**temp;
@@ -50,15 +49,14 @@ int	create_envcp_var(char *var, char *value, char **envcp, \
 	if (!newenvcp)
 		return (-1);
 	i = 0;
-	while (envcp[i])
+	while (data->envcp[i])
 		i++;
-	temp = (char **)realloc(envcp, sizeof(char *) * (i + 2));
+	temp = (char **)realloc(data->envcp, sizeof(char *) * (i + 2));
 	if (!temp)
 		return (free(newenvcp), -2);
-	program_data->envcp = temp;
-	envcp = temp;
-	envcp[i] = newenvcp;
-	envcp[i + 1] = NULL;
+	data->envcp = temp;
+	data->envcp[i] = newenvcp;
+	data->envcp[i + 1] = NULL;
 	return (0);
 }
 
@@ -66,20 +64,18 @@ int	create_envcp_var(char *var, char *value, char **envcp, \
 // if it fails to find the variable and createnew is 1, it creates a new one
 int	set_envcp_var(char *var, char *val, char createnew, t_program_data *data)
 {
-	int	i;
-	int	j;
-	int	k;
+	int			i;
+	int			k;
 
 	if (val == NULL)
 		val = "";
 	i = -1;
 	while (data->envcp[++i])
 	{
-		j = 0;
-		k = 0;
-		while (data->envcp[i][j] && var[k] && data->envcp[i][j++] == var[k++])
+		k = -1;
+		while (data->envcp[i][k++] && var[k] && data->envcp[i][k] == var[k])
 			;
-		if (var[k] == '\0' && data->envcp[i][j] == '=')
+		if (var[k] == '\0' && data->envcp[i][k] == '=')
 		{
 			free(data->envcp[i]);
 			data->envcp[i] = ft_strjoinfree(ft_strjoin(var, "="), \
@@ -89,8 +85,8 @@ int	set_envcp_var(char *var, char *val, char createnew, t_program_data *data)
 			return (0);
 		}
 	}
-	if (createnew)
-		return (create_envcp_var(var, val, data->envcp, data));
+	if (createnew == 1)
+		return (create_envcp_var(var, val, data));
 	return (-2);
 }