+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG minishell by nburchha & fschuber
 
-## v0.27.0 - 02.04.24 - Freddy
+## v0.28.0 - 03.04.24 - Freddy - 1670 / 2823
+- fixed export not working like bash bug for var-value pairs where var is only 1 char long
+- some fixes, and norminettification of envcp utils
+
+## v0.27.0 - 02.04.24 - Freddy - 1670 / 2823
 - moved builtins out of child processes
 	- made all of htem print to the outfd using ft_putstr_fd
 - fully removed wholly unnecessary cmd_start_index variable